Move publish status settings to separate card
