+++ cms/src/components/content/EditContentBasic.vue
@@ -1,15 +1,10 @@
 <script setup lang="ts">
 import LInput from "@/components/forms/LInput.vue";
-import LButton from "@/components/button/LButton.vue";
 import LCard from "@/components/common/LCard.vue";
-import FormLabel from "@/components/forms/FormLabel.vue";
 import { PencilIcon } from "@heroicons/vue/16/solid";
 import { PublishStatus, type ContentDto, db } from "luminary-shared";
-import { computed, nextTick, ref, watch } from "vue";
-import { DateTime } from "luxon";
+import { nextTick, ref, watch } from "vue";
 import { Slug } from "@/util/slug";
-import { BackspaceIcon } from "@heroicons/vue/20/solid";
-// import LToggle from "../forms/LToggle.vue";
 
 type Props = {
     disabled: boolean;
@@ -79,106 +74,6 @@ const validateSlug = async () => {
     if (!content.value) return;
     content.value.slug = await Slug.generate(content.value.slug, content.value._id || "");
 };
-
-// Publish and expiry dates
-const publishDateString = computed({
-    get() {
-        if (!content.value || !content.value.publishDate) return;
-        return db.toIsoDateTime(content.value.publishDate) || undefined;
-    },
-    set(val) {
-        if (!content.value) return;
-        if (!val) {
-            content.value.publishDate = undefined;
-            return;
-        }
-        content.value.publishDate = db.fromIsoDateTime(val);
-    },
-});
-
-const expiryDateString = computed({
-    get() {
-        if (!content.value || !content.value.expiryDate) return;
-        return db.toIsoDateTime(content.value.expiryDate) || undefined;
-    },
-    set(val) {
-        if (!content.value) return;
-        if (!val) {
-            content.value.expiryDate = undefined;
-            return;
-        }
-        content.value.expiryDate = db.fromIsoDateTime(val);
-    },
-});
-
-const selectedExpiryNumber = ref<number | undefined>(undefined);
-const selectedExpiryUnit = ref<string | undefined>(undefined);
-
-const calculateExpiryDate = () => {
-    if (
-        !content.value ||
-        !content.value.publishDate ||
-        !selectedExpiryNumber.value ||
-        !selectedExpiryUnit.value
-    )
-        return;
-
-    switch (selectedExpiryUnit.value) {
-        case "Week":
-            content.value.expiryDate = DateTime.fromMillis(content.value.publishDate)
-                .plus({ weeks: selectedExpiryNumber.value })
-                .toMillis();
-            break;
-        case "Month":
-            content.value.expiryDate = DateTime.fromMillis(content.value.publishDate)
-                .plus({ months: selectedExpiryNumber.value })
-                .toMillis();
-            break;
-        case "Year":
-            content.value.expiryDate = DateTime.fromMillis(content.value.publishDate)
-                .plus({ years: selectedExpiryNumber.value })
-                .toMillis();
-            break;
-        default:
-            console.warn(`Unknown unit: ${selectedExpiryUnit.value}`);
-    }
-    clearExpirySelection();
-};
-
-const setExpiryNumber = (number: number | undefined) => {
-    selectedExpiryNumber.value = number;
-    calculateExpiryDate();
-};
-
-const setExpiryUnit = (unit: string | undefined) => {
-    selectedExpiryUnit.value = unit;
-    calculateExpiryDate();
-};
-
-const clearExpirySelection = () => {
-    selectedExpiryNumber.value = undefined;
-    selectedExpiryUnit.value = undefined;
-};
-
-const clearExpiryDate = () => {
-    if (content.value) content.value.expiryDate = undefined;
-    clearExpirySelection();
-};
-
-// Linked publish and expiry dates
-// const linkedDates = ref<boolean>(false); // future feature
-
-// Publish status
-const publishStatus = computed({
-    get() {
-        if (!content.value) return false;
-        return content.value.status == PublishStatus.Published;
-    },
-    set(val) {
-        if (!content.value) return;
-        content.value.status = val ? PublishStatus.Published : PublishStatus.Draft;
-    },
-});
 </script>
 
 <template>
@@ -234,172 +129,5 @@ const publishStatus = computed({
             :disabled="disabled"
             v-model="content.summary"
         />
-
-        <!-- Publish settings -->
-        <div class="mt-4 flex flex-col gap-4 sm:flex-row">
-            <!-- Publish date -->
-            <LInput
-                name="publishDate"
-                label="Publish date"
-                class="sm:w-1/2"
-                type="datetime-local"
-                :disabled="disabled"
-                v-model="publishDateString"
-            >
-                <!-- Link publish & expiry dates toggle -->
-                <!-- <div class="flex items-center justify-between">
-                    <FormLabel>Link dates (not implemented yet)</FormLabel>
-                    <LToggle v-model="linkedDates" :disabled="disabled || true" />
-                </div> -->
-
-                <!-- Publish / draft toggle -->
-                <div class="mt-2 flex flex-col">
-                    <FormLabel>Content status:</FormLabel>
-                    <div class="mt-2 flex flex-col">
-                        <label
-                            class="text-md"
-                            :class="{ 'font-bold uppercase': !publishStatus }"
-                            data-test="draft"
-                        >
-                            <input
-                                type="radio"
-                                name="draftStatus"
-                                :value="false"
-                                v-model="publishStatus"
-                                :disabled="disabled"
-                                :class="{
-                                    'text-red-500': !publishStatus,
-                                    'text-gray-700': publishStatus,
-                                }"
-                            />
-                            Draft
-                        </label>
-                        <label
-                            class="text-md"
-                            :class="{ 'font-bold uppercase': publishStatus }"
-                            data-test="publish"
-                        >
-                            <input
-                                type="radio"
-                                name="publishStatus"
-                                :value="true"
-                                v-model="publishStatus"
-                                :disabled="disabled"
-                                :class="{
-                                    'text-green-500': publishStatus,
-                                    'text-gray-700': !publishStatus,
-                                }"
-                            />
-                            Published
-                        </label>
-                    </div>
-                </div>
-            </LInput>
-
-            <!-- Expiry date -->
-            <LInput
-                name="expiryDate"
-                label="Expiry date"
-                class="group sm:w-1/2"
-                type="datetime-local"
-                :disabled="disabled"
-                v-model="expiryDateString"
-            >
-                <!-- Expiry date shortcut buttons -->
-                <div class="flex w-full cursor-pointer flex-wrap gap-1">
-                    <LButton
-                        type="button"
-                        name="1"
-                        variant="secondary"
-                        class="flex-1"
-                        :class="{
-                            '!bg-black !text-white': selectedExpiryNumber === 1,
-                        }"
-                        @click="setExpiryNumber(1)"
-                        :disabled="disabled"
-                    >
-                        1
-                    </LButton>
-                    <LButton
-                        type="button"
-                        name="2"
-                        variant="secondary"
-                        class="flex-1"
-                        :class="{ '!bg-black !text-white': selectedExpiryNumber === 2 }"
-                        @click="setExpiryNumber(2)"
-                        :disabled="disabled"
-                    >
-                        2
-                    </LButton>
-                    <LButton
-                        type="button"
-                        name="3"
-                        variant="secondary"
-                        class="flex-1"
-                        :class="{ '!bg-black !text-white': selectedExpiryNumber === 3 }"
-                        @click="setExpiryNumber(3)"
-                        :disabled="disabled"
-                    >
-                        3
-                    </LButton>
-                    <LButton
-                        type="button"
-                        name="6"
-                        variant="secondary"
-                        class="flex-1"
-                        size="lg"
-                        :class="{ '!bg-black !text-white': selectedExpiryNumber === 6 }"
-                        @click="setExpiryNumber(6)"
-                        :disabled="disabled"
-                    >
-                        6
-                    </LButton>
-                    <LButton
-                        type="button"
-                        name="W"
-                        variant="secondary"
-                        class="flex-1"
-                        :class="{ '!bg-black !text-white': selectedExpiryUnit === 'Week' }"
-                        @click="setExpiryUnit('Week')"
-                        data-test="W"
-                        :disabled="disabled"
-                    >
-                        W
-                    </LButton>
-                    <LButton
-                        type="button"
-                        name="M"
-                        variant="secondary"
-                        class="flex-1"
-                        :class="{ '!bg-black !text-white': selectedExpiryUnit === 'Month' }"
-                        @click="setExpiryUnit('Month')"
-                        :disabled="disabled"
-                    >
-                        M
-                    </LButton>
-                    <LButton
-                        type="button"
-                        name="Y"
-                        variant="secondary"
-                        class="flex-1"
-                        :class="{ '!bg-black !text-white': selectedExpiryUnit === 'Year' }"
-                        @click="setExpiryUnit('Year')"
-                        :disabled="disabled"
-                    >
-                        Y
-                    </LButton>
-                    <LButton
-                        type="button"
-                        name="clear"
-                        variant="secondary"
-                        :icon="BackspaceIcon"
-                        class="flex-1"
-                        @click="clearExpiryDate()"
-                        :disabled="disabled"
-                    >
-                    </LButton>
-                </div>
-            </LInput>
-        </div>
     </LCard>
 </template>
